requirements.txt
streamlit>=1.30
biopython>=1.83
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
